+++ src/main.jsx
@@ -2,8 +2,8 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { createBrowserRouter, RouterProvider, Outlet } from "react-router";
 import "./index.css";
-import LoggedInLayout from "./LoggedInLayout.jsx";
-import LoggedOutLayout from "./LoggedOutLayout.jsx";
+import LoggedInLayout from "./layouts/LoggedInLayout.jsx";
+import LoggedOutLayout from "./layouts/LoggedOutLayout.jsx";
 import Login from "./pages/Login.jsx";
 import SignUp from "./pages/SignUp.jsx";
 import Dashboard from "./pages/Dashboard.jsx";
